enhancement to rendering. admin_task enhancement
Updated the rendering engine so that the top-most element of the
template itself can now contain a data-bind that gets rendered. very
handy for processList (hiding a row based on some criteria, for
example).
added a new renderFormat to the task list to hide tasks that have been
completed for more than a week.